+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
+export { useClickOutside } from "@/hooks/useClickOutside";
 export { useCounter } from "@/hooks/useCounter";
 export type { UseCounterOptions } from "@/hooks/useCounter";
 export { useDragScroll } from "@/hooks/useDragScroll";
